Fixed borken import statements and removed an extra file.
